move code to kubernetes.go
